styled calendar component
