+++ client/src/containers/CreateTrip/CreateTrip.jsx
@@ -1,27 +1,12 @@
-import React, { useState } from "react";
-import DatePicker from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
+import React from "react";
+import Calendar from "../../components/Calendar/Calendar";
 
 const CreateTrip = () => {
-  const [startDate, setStartDate] = useState(new Date());
-  const [endDate, setEndDate] = useState(null);
-  const onChange = (dates) => {
-    const [start, end] = dates;
-    setStartDate(start);
-    setEndDate(end);
-  };
   return (
-    <>
-      <h1>This is the Create Trip Page</h1>
-      <DatePicker
-        selected={startDate}
-        onChange={onChange}
-        startDate={startDate}
-        endDate={endDate}
-        selectsRange
-        inline
-      />
-    </>
+    <div className="content has-text-centered">
+      <h1>Create a Trip!</h1>
+      <Calendar />
+    </div>
   );
 };
