+++ client/src/sass/App.scss
@@ -5,7 +5,7 @@
 
 // Set your brand colors
 $purple: #2E3560;
-$purple-lighter: #3C447C;
+$purple-lighter: #938DB9;
 $teal: #33C1BF;
 $teal-lighter: #8EE1E0;
 $orange: #F48E3F;
